refactor: Increase physics controller timestep precision
data type for times in PhysicsController::update() changed from float to long double

+++ src/Controller/PhysicsController.cpp
@@ -27,8 +27,8 @@ void PhysicsController::init() {
 }
 
 void PhysicsController::update() {
-    float currentTime = glfwGetTime();
-    float deltaTime = currentTime - previousTime;
+    long double currentTime = glfwGetTime();
+    long double deltaTime = currentTime - previousTime;
     previousTime = currentTime;
 
     accumulator += deltaTime;
